refactor(files): 拆分文件摄入为格式专属子模块
