+++ pinvou3-app/src-tauri/src/features/files/file_ingest.rs
@@ -13,10 +13,41 @@
 //! `warning: "需要安装 ..."`，前端 chip 显示，不阻塞其它格式。
 //!
 //! Token 估算：粗算 `chars / 1.6`（中英混合保守值）。不引 tiktoken-rs 减依赖。
+//!
+//! 本文件是 **facade**：只承载主入口 `ingest` 派发、共享类型 [`IngestResult`] /
+//! [`MAX_FILE_BYTES`]、路径校验工具与 placeholder。各格式摄入逻辑拆到子模块：
+//! - [`ingest_deps`]：系统工具探测 + 依赖体检 + 外部命令构建
+//! - [`text_decode`]：纯文本摄入 + 内容嗅探 / 解码
+//! - [`ingest_pdf`]：PDF 摄入
+//! - [`ingest_office`]：pandoc / LibreOffice / 表格 / 演示摄入
+//! - [`ingest_archive`]：压缩包摄入
+//! - [`ingest_email`]：邮件摄入
+//! - [`visual_preview`]：OCR + 可视化预览（base64 / data URI / 逐页 PNG）
+
+#[path = "ingest_archive.rs"]
+mod ingest_archive;
+#[path = "ingest_deps.rs"]
+mod ingest_deps;
+#[path = "ingest_email.rs"]
+mod ingest_email;
+#[path = "ingest_office.rs"]
+mod ingest_office;
+#[path = "ingest_pdf.rs"]
+mod ingest_pdf;
+#[path = "text_decode.rs"]
+mod text_decode;
+#[path = "visual_preview.rs"]
+mod visual_preview;
+
+pub use ingest_deps::{
+    check_dependencies, install_dependencies, system_tools, DependencyCheckItem, SystemTools,
+};
+pub use visual_preview::{
+    base64_encode, image_file_to_data_uri, libreoffice_to_inline_html, ocr_image_for_kb,
+    office_to_png_data_uris, pdf_to_png_data_uris,
+};
 
 use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::sync::OnceLock;
 
 use serde::{Deserialize, Serialize};
 
@@ -106,213 +137,11 @@ impl IngestResult {
     }
 }
 
-#[derive(Debug, Clone, Copy, Default, Serialize)]
-pub struct SystemTools {
-    pub pandoc: bool,
-    pub pdftotext: bool,
-    /// LibreOffice headless —— 用来转 .doc / .ppt / .xls 等旧 office 格式。
-    pub libreoffice: bool,
-    /// tesseract OCR —— 图片文字识别 + 扫描件 PDF 兜底。
-    pub tesseract: bool,
-    /// pdftoppm（poppler-utils）—— 把扫描件 PDF 逐页转图再喂给 tesseract。
-    pub pdftoppm: bool,
-    /// 7z —— 解 zip/7z/rar 等压缩包；Windows 优先使用内置 7-Zip。
-    pub sevenzip: bool,
-    /// Python —— 解析 .eml 邮件（标准库 email 模块，无额外依赖）。
-    pub python: bool,
-    /// msgconvert（libemail-outlook-message-perl）—— Linux .msg → .eml；Windows 走 Rust 原生解析。
-    pub msgconvert: bool,
-}
-
-static SYSTEM_TOOLS: OnceLock<SystemTools> = OnceLock::new();
-
-/// 启动时（或第一次 ingest 时）检测一次系统工具。
-pub fn system_tools() -> SystemTools {
-    *SYSTEM_TOOLS.get_or_init(|| SystemTools {
-        pandoc: crate::platform::os::pandoc_tool_exists(),
-        pdftotext: crate::platform::os::pdf_tool_exists("pdftotext"),
-        libreoffice: crate::platform::os::command_exists("soffice")
-            || crate::platform::os::command_exists("libreoffice"),
-        tesseract: crate::platform::os::ocr_tool_exists(),
-        pdftoppm: crate::platform::os::pdf_tool_exists("pdftoppm"),
-        sevenzip: crate::platform::os::archive_tool_exists(),
-        python: crate::platform::os::command_exists(&crate::platform::paths::python_command()),
-        msgconvert: !crate::platform::os::msg_converter_required()
-            || crate::platform::os::command_exists("msgconvert"),
-    })
-}
-
-/// 设置页「依赖体检」一项：一类文件解析能力 + 它所需系统工具是否齐全 + 缺失时的 apt 包。
-/// `apt` 是空格分隔的包名串，可直接拼进 `sudo apt install <apt>`。能力名走前端 i18n（按 key 映射）。
-/// `hint` 为缺省时给用户的手动安装指引（如 macOS 邮件依赖无 Homebrew formula），
-/// 前端优先于 `apt` 显示；`skip_serializing_if` 保证老前端无感知。
-#[derive(Debug, Clone, Serialize)]
-pub struct DependencyCheckItem {
-    pub key: String,
-    pub installed: bool,
-    pub apt: String,
-    pub install_action: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hint: Option<String>,
-}
-
-/// 一键安装进度的实时事件载荷。平台适配器经纯 Rust 回调上报,
-/// 本域层包成此结构后 `app.emit("deps:install_progress", …)`。
-#[derive(Clone, Serialize)]
-pub struct DepInstallProgress {
-    pub package: String,
-    pub current: usize,
-    pub total: usize,
-    pub detail: Option<String>,
-}
-
-/// 体检各项可选依赖的安装状态。**实时检测（不走 `system_tools` 的 OnceLock 缓存）**——
-/// 用户照提示装完依赖后重新体检要能立刻反映，不能被首次缓存钉死。命名/分组与
-/// `ingest` 内各格式分支的 warning 文案同源，缺啥装啥一致。
-pub fn check_dependencies() -> Vec<DependencyCheckItem> {
-    let item = |key: &str, installed: bool, apt: &str| DependencyCheckItem {
-        key: key.into(),
-        installed,
-        apt: apt.into(),
-        install_action: None,
-        hint: None,
-    };
-    let libreoffice = crate::platform::os::command_exists("soffice")
-        || crate::platform::os::command_exists("libreoffice");
-    let mut items = Vec::new();
-    if crate::platform::os::show_pdf_dependency_check() {
-        items.push(item(
-            "pdf",
-            crate::platform::os::pdf_tool_exists("pdftotext"),
-            crate::platform::os::pdf_dependency_packages(),
-        ));
-    }
-    if crate::platform::os::show_pandoc_dependency_check() {
-        items.push(item(
-            "office_modern",
-            crate::platform::os::pandoc_tool_exists(),
-            crate::platform::os::pandoc_dependency_packages(),
-        ));
-    }
-    items.push(item(
-        "voice_asr",
-        crate::features::voice::asr_tool_exists(),
-        crate::features::voice::asr_dependency_packages(),
-    ));
-    items.push(item("office_legacy", libreoffice, "libreoffice"));
-    if crate::platform::os::show_ocr_dependency_check() {
-        items.push(item(
-            "ocr",
-            crate::platform::os::ocr_tool_exists()
-                && crate::platform::os::pdf_tool_exists("pdftoppm"),
-            crate::platform::os::ocr_dependency_packages(),
-        ));
-    }
-    if crate::platform::os::show_archive_dependency_check() {
-        items.push(item(
-            "archive",
-            crate::platform::os::archive_tool_exists(),
-            crate::platform::os::archive_dependency_packages(),
-        ));
-    }
-    // 邮件依赖(msgconvert)来自 Perl 模块 Email::Outlook::Message,无 Homebrew formula,
-    // 无法一键安装。给出手动安装指引(hint),前端优先于 apt 显示,让用户知道怎么装;
-    // email_dependency_packages() 仍返回空串,确保不参与一键安装。
-    let email_hint = crate::platform::os::email_manual_hint().map(str::to_string);
-    items.push(DependencyCheckItem {
-        key: "email".into(),
-        installed: crate::platform::os::email_tool_exists(),
-        apt: crate::platform::os::email_dependency_packages().into(),
-        install_action: None,
-        hint: email_hint,
-    });
-    items
-}
-
-fn pdf_tool_command(command: &str) -> Command {
-    crate::platform::process::HiddenCommand::new(crate::platform::os::pdf_tool_path(command))
-}
-
-fn pandoc_tool_command() -> Command {
-    crate::platform::process::HiddenCommand::new(crate::platform::os::pandoc_tool_path())
-}
-
-fn ocr_tool_command() -> Command {
-    crate::platform::process::HiddenCommand::new(crate::platform::os::ocr_tool_path())
-}
-
-fn archive_tool_command() -> Command {
-    crate::platform::process::HiddenCommand::new(crate::platform::os::archive_tool_path())
-}
-
-fn libreoffice_tool_command() -> Command {
-    crate::platform::process::HiddenCommand::new(crate::platform::os::libreoffice_tool_path())
-}
-
-fn libreoffice_user_installation_arg(profile_dir: &Path) -> Result<String, String> {
-    Ok(format!(
-        "-env:UserInstallation={}",
-        crate::platform::os::file_url_from_path(profile_dir)?
-    ))
-}
-
-fn add_ocr_tessdata_arg(command: &mut Command) {
-    if let Some(tessdata_dir) = crate::platform::os::ocr_tessdata_dir() {
-        command.arg("--tessdata-dir").arg(tessdata_dir);
-    }
-}
-
-/// 体检卡「一键安装」：委托 OS 调度层安装缺失依赖。
-/// Linux 由 OS 层保留包名白名单和 pkexec/apt 行为；其他系统清晰降级。
-/// `app` 用于把平台适配器上报的纯 Rust 进度回调转成 Tauri 事件
-/// `deps:install_progress`,前端据此实时刷新「正在安装 X (n/总数)…」,
-/// 不再全程只有静态「安装中…」(libreoffice cask 长尾尤其像卡死)。
-pub async fn install_dependencies(
-    app: tauri::AppHandle,
-    packages: Vec<String>,
-) -> Result<(), String> {
-    use tauri::Emitter;
-    tokio::task::spawn_blocking(move || {
-        // AppHandle 是 Clone+Send,克隆一份进阻塞线程(既有先例 dingtalk.rs
-        // spawn_blocking(move || run_connect_flow(&app2)))。
-        let app = app.clone();
-        let report = move |package: &str, current: usize, total: usize, detail: Option<&str>| {
-            let _ = app.emit(
-                "deps:install_progress",
-                DepInstallProgress {
-                    package: package.to_string(),
-                    current,
-                    total,
-                    detail: detail.map(str::to_string),
-                },
-            );
-        };
-        crate::features::dependencies::install_dependencies(packages, Some(&report))
-    })
-    .await
-    .map_err(|e| format!("安装任务失败: {e}"))?
-}
-
-/// tesseract 的 `-l` 语言参数。pinvou3 面向国内政企，中文是刚需，所以优先
-/// `chi_sim+eng`；若没装中文包(`tesseract-ocr-chi-sim`)则降级 `eng`，不报错。
-/// 探测一次缓存：跑 `tesseract --list-langs` 看输出里有没有 `chi_sim`。
-fn ocr_lang_arg() -> String {
-    static LANG: OnceLock<String> = OnceLock::new();
-    LANG.get_or_init(|| {
-        let mut command = ocr_tool_command();
-        command.arg("--list-langs");
-        add_ocr_tessdata_arg(&mut command);
-        let listed = command
-            .output()
-            .map(|o| String::from_utf8_lossy(&o.stdout).into_owned())
-            .unwrap_or_default();
-        if listed.lines().any(|l| l.trim() == "chi_sim") {
-            "chi_sim+eng".to_string()
-        } else {
-            "eng".to_string()
-        }
-    })
-    .clone()
+/// 粗算 token：中英混合按 `chars / 1.6` —— 比较保守，偏向高估避免炸上下文。
+/// 实测 cl100k_base 中文 1.0-1.5 char/token、英文 3-4 char/token。
+fn estimate_tokens(text: &str) -> u32 {
+    let chars = text.chars().count();
+    (chars as f64 / 1.6).ceil() as u32
 }
 
 /// 主入口：派发到不同处理函数，返回统一 IngestResult。
@@ -357,25 +186,27 @@ pub fn ingest(path: &Path) -> IngestResult {
     let kind = classify(&ext);
 
     match kind {
-        "text" => ingest_text(path, basename, path_str, byte_size),
-        "pdf" => ingest_pdf(path, basename, path_str, byte_size),
+        "text" => text_decode::ingest_text(path, basename, path_str, byte_size),
+        "pdf" => ingest_pdf::ingest_pdf(path, basename, path_str, byte_size),
         // 文字文档：pandoc 原生支持 docx/odt。
-        "doc_pandoc" => ingest_pandoc(path, basename, byte_size, &ext),
+        "doc_pandoc" => ingest_office::ingest_pandoc(path, basename, byte_size, &ext),
         // 文字文档：pandoc 吃不下，LibreOffice 转纯文本（doc/rtf/wps）。
-        "doc_office" => ingest_office_text(path, basename, byte_size, &ext),
+        "doc_office" => ingest_office::ingest_office_text(path, basename, byte_size, &ext),
         // 演示：LibreOffice 转 PDF 再 pdftotext（pptx/ppt/odp/dps）。
-        "presentation" => ingest_presentation(path, basename, byte_size, &ext),
+        "presentation" => ingest_office::ingest_presentation(path, basename, byte_size, &ext),
         // 表格：LibreOffice 转 CSV（xlsx/ods/xls/et）—— pandoc 不支持表格输入。
-        "spreadsheet" => ingest_spreadsheet(path, basename, path_str, byte_size, &ext),
-        "image" => ingest_image(path, basename, path_str, byte_size),
-        "archive" => ingest_archive(path, basename, path_str, byte_size),
-        "email" => ingest_email(path, basename, path_str, byte_size, &ext),
+        "spreadsheet" => {
+            ingest_office::ingest_spreadsheet(path, basename, path_str, byte_size, &ext)
+        }
+        "image" => visual_preview::ingest_image(path, basename, path_str, byte_size),
+        "archive" => ingest_archive::ingest_archive(path, basename, path_str, byte_size),
+        "email" => ingest_email::ingest_email(path, basename, path_str, byte_size, &ext),
         "media" => media_placeholder(basename, path_str, byte_size),
         // 私钥 / 密钥库:绝不读正文(防泄露给 LLM)。见 classify 的 "secret" 分支。
         "secret" => secret_placeholder(basename, path_str, byte_size),
         // 未知扩展名 / 无扩展名:不再一刀切 binary,先内容嗅探。是文本就当文本读
         // (让模型看到内容 —— 这是「接受任何文件」的根本解),真二进制才降级。
-        _ => sniff_text_or_binary(path, basename, path_str, byte_size),
+        _ => text_decode::sniff_text_or_binary(path, basename, path_str, byte_size),
     }
 }
 
@@ -436,1434 +267,6 @@ fn classify(ext: &str) -> &'static str {
     }
 }
 
-fn ingest_text(path: &Path, basename: String, path_str: String, byte_size: u64) -> IngestResult {
-    // 先读字节:严格 UTF-8 直接用;非法 UTF-8(如 GBK 中文 .txt)lossy 还原并标 warning,
-    // 而非返回空 markdown —— 中文用户常有非 UTF-8 文本文件,lossy 能保住大部分可读内容。
-    // 读字节本身失败才回「读取失败」。
-    let bytes = match std::fs::read(path) {
-        Ok(b) => b,
-        Err(e) => {
-            return IngestResult::warning(
-                "text",
-                &basename,
-                path,
-                byte_size,
-                format!("读取失败(可能不是文本): {e}"),
-            );
-        }
-    };
-    // 内容侧安全兜底:已知文本扩展名也可能被塞了私钥(如把 id_rsa 改名 id_rsa.txt),
-    // 这种情况按 secret 拒绝读取,不进 markdown。
-    if looks_like_secret_material(&bytes) {
-        return secret_placeholder(basename, path_str, byte_size);
-    }
-    let (content, warning) = decode_text_bytes(&bytes);
-    let tokens = content.as_ref().map(|c| estimate_tokens(c)).unwrap_or(0);
-    IngestResult {
-        kind: "text".into(),
-        basename,
-        path: path_str,
-        markdown: content,
-        token_estimate: tokens,
-        byte_size,
-        warning,
-    }
-}
-
-fn ingest_pdf(path: &Path, basename: String, path_str: String, byte_size: u64) -> IngestResult {
-    let tools = system_tools();
-    if !tools.pdftotext {
-        return IngestResult::warning(
-            "pdf",
-            &basename,
-            path,
-            byte_size,
-            crate::platform::os::pdf_text_missing_message(),
-        );
-    }
-    // pdftotext -layout <path> -  → stdout
-    let out = pdf_tool_command("pdftotext")
-        .arg("-layout")
-        .arg(path)
-        .arg("-")
-        .output();
-    match out {
-        Ok(o) if o.status.success() => {
-            let content = String::from_utf8_lossy(&o.stdout).into_owned();
-            // 扫描件 PDF（图层 PDF）没有文字层，pdftotext 返回空白 —— 此时
-            // 走 OCR 兜底：pdftoppm 逐页转图再喂 tesseract。
-            if content.trim().is_empty() {
-                return ocr_pdf(path, basename, path_str, byte_size);
-            }
-            IngestResult::with_markdown("pdf", &basename, path, byte_size, content)
-        }
-        Ok(o) => IngestResult::warning(
-            "pdf",
-            &basename,
-            path,
-            byte_size,
-            format!(
-                "pdftotext 失败: {}",
-                String::from_utf8_lossy(&o.stderr).trim()
-            ),
-        ),
-        Err(e) => IngestResult::warning(
-            "pdf",
-            &basename,
-            path,
-            byte_size,
-            format!("pdftotext 调用失败: {e}"),
-        ),
-    }
-}
-
-fn ingest_pandoc(path: &Path, basename: String, byte_size: u64, label: &str) -> IngestResult {
-    let tools = system_tools();
-    if !tools.pandoc {
-        return IngestResult::warning(
-            label,
-            &basename,
-            path,
-            byte_size,
-            crate::platform::os::pandoc_missing_message(),
-        );
-    }
-    let out = pandoc_tool_command()
-        .arg("-t")
-        .arg("markdown")
-        .arg(path)
-        .output();
-    match out {
-        Ok(o) if o.status.success() => {
-            let content = String::from_utf8_lossy(&o.stdout).into_owned();
-            IngestResult::with_markdown(label, &basename, path, byte_size, content)
-        }
-        Ok(o) => IngestResult::warning(
-            label,
-            &basename,
-            path,
-            byte_size,
-            format!("pandoc 失败: {}", String::from_utf8_lossy(&o.stderr).trim()),
-        ),
-        Err(e) => IngestResult::warning(
-            label,
-            &basename,
-            path,
-            byte_size,
-            format!("pandoc 调用失败: {e}"),
-        ),
-    }
-}
-
-/// 用 LibreOffice headless 把文件转成指定 filter 的产物并读回文本。复用于旧
-/// office、WPS 文字、电子表格等 pandoc 吃不下的格式。`convert_to` 是 soffice 的
-/// 输出 filter 串，`out_ext` 是产物扩展名（用来在临时目录里定位输出文件）。
-fn libreoffice_convert_text(
-    path: &Path,
-    convert_to: &str,
-    out_ext: &str,
-) -> Result<String, String> {
-    if !system_tools().libreoffice {
-        return Err(crate::platform::os::libreoffice_missing_message().into());
-    }
-    // 临时目录：每次唯一，避免并发文件名冲突。
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-libreoffice-{ts}"));
-    std::fs::create_dir_all(&tmpdir).map_err(|e| format!("创建临时目录失败: {e}"))?;
-
-    // 独立 UserInstallation profile：LibreOffice 同一 profile 不能并发(会 lock)，
-    // 用户一次拖多个 office 文件时前端会并发 ingest_file，必须各用各的 profile。
-    let out = libreoffice_tool_command()
-        .arg(libreoffice_user_installation_arg(&tmpdir.join("profile"))?)
-        .arg("--headless")
-        .arg("--convert-to")
-        .arg(convert_to)
-        .arg("--outdir")
-        .arg(&tmpdir)
-        .arg(path)
-        .output();
-
-    let result = match out {
-        Ok(o) if o.status.success() => {
-            let stem = path
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .unwrap_or("converted");
-            let out_path = tmpdir.join(format!("{stem}.{out_ext}"));
-            std::fs::read_to_string(&out_path)
-                // soffice 的 txt/csv 导出会带 UTF-8 BOM，去掉以免污染正文开头。
-                .map(|s| s.trim_start_matches('\u{feff}').to_string())
-                .map_err(|e| format!("LibreOffice 转换后读取失败: {e}"))
-        }
-        Ok(o) => Err(format!(
-            "LibreOffice 转换失败: {}",
-            String::from_utf8_lossy(&o.stderr).trim()
-        )),
-        Err(e) => Err(format!("LibreOffice 调用失败: {e}")),
-    };
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-// ============== 产物可视化预览助手（commands::render_artifact_visual 复用）==============
-
-/// 极简 base64 标准编码（无换行）。仅为内联图片 data URI 用，不值得引第三方 crate。
-pub fn base64_encode(data: &[u8]) -> String {
-    const TABLE: &[u8; 64] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    let mut out = String::with_capacity(data.len().div_ceil(3) * 4);
-    for chunk in data.chunks(3) {
-        let b0 = chunk[0] as u32;
-        let b1 = *chunk.get(1).unwrap_or(&0) as u32;
-        let b2 = *chunk.get(2).unwrap_or(&0) as u32;
-        let n = (b0 << 16) | (b1 << 8) | b2;
-        out.push(TABLE[((n >> 18) & 63) as usize] as char);
-        out.push(TABLE[((n >> 12) & 63) as usize] as char);
-        out.push(if chunk.len() > 1 {
-            TABLE[((n >> 6) & 63) as usize] as char
-        } else {
-            '='
-        });
-        out.push(if chunk.len() > 2 {
-            TABLE[(n & 63) as usize] as char
-        } else {
-            '='
-        });
-    }
-    out
-}
-
-/// 图片扩展名 → MIME。用于 data URI 前缀。
-fn image_mime(ext: &str) -> &'static str {
-    match ext.to_ascii_lowercase().as_str() {
-        "png" => "image/png",
-        "jpg" | "jpeg" => "image/jpeg",
-        "gif" => "image/gif",
-        "webp" => "image/webp",
-        "bmp" => "image/bmp",
-        "svg" => "image/svg+xml",
-        _ => "application/octet-stream",
-    }
-}
-
-/// 单个图片文件 → `data:image/...;base64,...`。
-pub fn image_file_to_data_uri(path: &Path) -> Result<String, String> {
-    let bytes = std::fs::read(path).map_err(|e| format!("读图失败: {e}"))?;
-    let ext = path.extension().and_then(|e| e.to_str()).unwrap_or("");
-    Ok(format!(
-        "data:{};base64,{}",
-        image_mime(ext),
-        base64_encode(&bytes)
-    ))
-}
-
-/// 把 HTML 里指向本地旁置图片的 `src` 引用 base64 内联,产出自包含 HTML。
-/// soffice 导出的 HTML 把图片写成 `<stem>_html_xxx.png` 同目录文件,iframe `srcDoc`
-/// 加载不到这些相对路径 → 全部内联。已是 data:/http 的跳过。双引号、单引号都处理。
-fn inline_html_images(html: &str, dir: &Path) -> String {
-    let mut result = html.to_string();
-    for quote in ['"', '\''] {
-        let needle = format!("src={quote}");
-        let mut rebuilt = String::with_capacity(result.len());
-        let mut from = 0;
-        loop {
-            match result[from..].find(&needle) {
-                Some(rel) => {
-                    let val_start = from + rel + needle.len();
-                    match result[val_start..].find(quote) {
-                        Some(endrel) => {
-                            let val_end = val_start + endrel;
-                            let val = &result[val_start..val_end];
-                            rebuilt.push_str(&result[from..val_start]); // 含 src="
-                            if val.starts_with("data:") || val.starts_with("http") || val.is_empty()
-                            {
-                                rebuilt.push_str(val);
-                            } else {
-                                let fname = val.trim_start_matches("./");
-                                match image_file_to_data_uri(&dir.join(fname)) {
-                                    Ok(uri) => rebuilt.push_str(&uri),
-                                    Err(_) => rebuilt.push_str(val),
-                                }
-                            }
-                            from = val_end; // 闭合引号留给下一轮拼接
-                        }
-                        None => {
-                            rebuilt.push_str(&result[from..]);
-                            break;
-                        }
-                    }
-                }
-                None => {
-                    rebuilt.push_str(&result[from..]);
-                    break;
-                }
-            }
-        }
-        result = rebuilt;
-    }
-    result
-}
-
-/// office 文档 → 可视化 HTML（版式/图片还原）。soffice `--convert-to html`,旁置图片
-/// 内联成自包含 HTML 返回,前端直接喂 iframe srcDoc。复用 libreoffice_convert_text 的
-/// 独立 UserInstallation profile + 临时目录约定。
-pub fn libreoffice_to_inline_html(path: &Path) -> Result<String, String> {
-    if !system_tools().libreoffice {
-        return Err(crate::platform::os::libreoffice_missing_message().into());
-    }
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-lo-html-{ts}"));
-    std::fs::create_dir_all(&tmpdir).map_err(|e| format!("创建临时目录失败: {e}"))?;
-
-    let out = libreoffice_tool_command()
-        .arg(libreoffice_user_installation_arg(&tmpdir.join("profile"))?)
-        .arg("--headless")
-        .arg("--convert-to")
-        // 不写死 `html:HTML`(那是 Writer 专用 filter,套到 Calc/Impress 会无产出)。
-        // 只给 `html` → LibreOffice 按文档类型自动选对应 HTML 导出 filter。
-        .arg("html")
-        .arg("--outdir")
-        .arg(&tmpdir)
-        .arg(path)
-        .output();
-
-    let result = (|| -> Result<String, String> {
-        match out {
-            Ok(o) if o.status.success() => {}
-            Ok(o) => {
-                return Err(format!(
-                    "LibreOffice 转换失败: {}",
-                    String::from_utf8_lossy(&o.stderr).trim()
-                ))
-            }
-            Err(e) => return Err(format!("LibreOffice 调用失败: {e}")),
-        }
-        // 不假设产物叫 `<stem>.html`(filter 不同 / 文件名带特殊字符都可能变)——
-        // 扫临时目录里产出的 .html(优先匹配 stem,否则取首个)。
-        let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
-        let htmls: Vec<PathBuf> = std::fs::read_dir(&tmpdir)
-            .map(|rd| {
-                rd.filter_map(|e| e.ok().map(|e| e.path()))
-                    .filter(|p| {
-                        matches!(
-                            p.extension().and_then(|e| e.to_str()),
-                            Some("html") | Some("htm")
-                        )
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-        let pick = htmls
-            .iter()
-            .find(|p| p.file_stem().and_then(|s| s.to_str()) == Some(stem))
-            .or_else(|| htmls.first())
-            .ok_or_else(|| "LibreOffice 未产出 HTML".to_string())?;
-        let html = std::fs::read_to_string(pick).map_err(|e| format!("读取转换 HTML 失败: {e}"))?;
-        Ok(inline_html_images(&html, &tmpdir))
-    })();
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-/// 演示稿(pptx/ppt/odp)→ 先转 PDF 再逐页转 PNG。Impress 的 HTML 导出会拆成一堆
-/// 文件且版式失真,转 PDF→PNG 更可靠:每页 = 一张幻灯片。复用 110 dpi。
-pub fn office_to_png_data_uris(path: &Path, max_pages: u32) -> Result<(Vec<String>, bool), String> {
-    let tools = system_tools();
-    if !tools.libreoffice {
-        return Err(crate::platform::os::libreoffice_missing_message().into());
-    }
-    if !tools.pdftoppm {
-        return Err(crate::platform::os::pdf_render_missing_message().into());
-    }
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-office-png-{ts}"));
-    std::fs::create_dir_all(&tmpdir).map_err(|e| format!("创建临时目录失败: {e}"))?;
-
-    let result = (|| -> Result<(Vec<String>, bool), String> {
-        // 1) office → PDF
-        let out = libreoffice_tool_command()
-            .arg(libreoffice_user_installation_arg(&tmpdir.join("profile"))?)
-            .arg("--headless")
-            .arg("--convert-to")
-            .arg("pdf")
-            .arg("--outdir")
-            .arg(&tmpdir)
-            .arg(path)
-            .output();
-        match out {
-            Ok(o) if o.status.success() => {}
-            Ok(o) => {
-                return Err(format!(
-                    "LibreOffice 转 PDF 失败: {}",
-                    String::from_utf8_lossy(&o.stderr).trim()
-                ))
-            }
-            Err(e) => return Err(format!("LibreOffice 调用失败: {e}")),
-        }
-        // 找产出的 PDF(扫目录,别假设文件名)。
-        let pdf = std::fs::read_dir(&tmpdir)
-            .ok()
-            .and_then(|rd| {
-                rd.filter_map(|e| e.ok().map(|e| e.path()))
-                    .find(|p| p.extension().and_then(|e| e.to_str()) == Some("pdf"))
-            })
-            .ok_or_else(|| "LibreOffice 未产出 PDF".to_string())?;
-
-        // 2) PDF → PNG 页(直接在同一 tmpdir,避免再开目录)。
-        let prefix = tmpdir.join("page");
-        let conv = pdf_tool_command("pdftoppm")
-            .arg("-png")
-            .arg("-r")
-            .arg("110")
-            .arg("-l")
-            .arg(max_pages.to_string())
-            .arg(&pdf)
-            .arg(&prefix)
-            .output();
-        match conv {
-            Ok(o) if o.status.success() => {}
-            Ok(o) => {
-                return Err(format!(
-                    "pdftoppm 转图失败: {}",
-                    String::from_utf8_lossy(&o.stderr).trim()
-                ))
-            }
-            Err(e) => return Err(format!("pdftoppm 调用失败: {e}")),
-        }
-        let mut pages: Vec<PathBuf> = std::fs::read_dir(&tmpdir)
-            .map(|rd| {
-                rd.filter_map(|e| e.ok().map(|e| e.path()))
-                    .filter(|p| {
-                        p.extension().and_then(|e| e.to_str()) == Some("png")
-                            && p.file_stem()
-                                .and_then(|s| s.to_str())
-                                .map(|s| s.starts_with("page"))
-                                .unwrap_or(false)
-                    })
-                    .collect()
-            })
-            .unwrap_or_default();
-        pages.sort();
-        if pages.is_empty() {
-            return Err("未产出可渲染幻灯片页".into());
-        }
-        let truncated = pages.len() as u32 >= max_pages;
-        let mut uris = Vec::with_capacity(pages.len());
-        for p in &pages {
-            uris.push(image_file_to_data_uri(p)?);
-        }
-        Ok((uris, truncated))
-    })();
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-/// PDF → 逐页 PNG 的 data URI 列表(可视化预览)。复用 ocr_pdf 的 pdftoppm 调用样板,
-/// 但 110 dpi(预览够清又不至于 data URI 过大)。返回 (data_uris, 是否因上限截断)。
-pub fn pdf_to_png_data_uris(path: &Path, max_pages: u32) -> Result<(Vec<String>, bool), String> {
-    if !system_tools().pdftoppm {
-        return Err(crate::platform::os::pdf_render_missing_message().into());
-    }
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-pdfpreview-{ts}"));
-    std::fs::create_dir_all(&tmpdir).map_err(|e| format!("创建临时目录失败: {e}"))?;
-    let prefix = tmpdir.join("page");
-
-    let convert = pdf_tool_command("pdftoppm")
-        .arg("-png")
-        .arg("-r")
-        .arg("110")
-        .arg("-l")
-        .arg(max_pages.to_string())
-        .arg(path)
-        .arg(&prefix)
-        .output();
-
-    let result = (|| -> Result<(Vec<String>, bool), String> {
-        match convert {
-            Ok(o) if o.status.success() => {}
-            Ok(o) => {
-                return Err(format!(
-                    "pdftoppm 转图失败: {}",
-                    String::from_utf8_lossy(&o.stderr).trim()
-                ))
-            }
-            Err(e) => return Err(format!("pdftoppm 调用失败: {e}")),
-        }
-        let mut pages: Vec<PathBuf> = std::fs::read_dir(&tmpdir)
-            .map(|rd| {
-                rd.filter_map(|e| e.ok().map(|e| e.path()))
-                    .filter(|p| p.extension().and_then(|e| e.to_str()) == Some("png"))
-                    .collect()
-            })
-            .unwrap_or_default();
-        pages.sort();
-        if pages.is_empty() {
-            return Err("PDF 未产出可渲染页".into());
-        }
-        let truncated = pages.len() as u32 >= max_pages;
-        let mut uris = Vec::with_capacity(pages.len());
-        for p in &pages {
-            uris.push(image_file_to_data_uri(p)?);
-        }
-        Ok((uris, truncated))
-    })();
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-/// 文字类文档（.doc/.rtf + WPS .wps）：pandoc 吃不下，用 LibreOffice 转纯文本。
-fn ingest_office_text(path: &Path, basename: String, byte_size: u64, kind: &str) -> IngestResult {
-    match libreoffice_convert_text(path, "txt:Text (encoded):UTF8", "txt") {
-        Ok(content) => IngestResult::with_markdown(kind, &basename, path, byte_size, content),
-        Err(e) => IngestResult::warning(kind, &basename, path, byte_size, e),
-    }
-}
-
-/// 电子表格（.xlsx/.ods/.xls + WPS .et）：用 calamine 纯 Rust 读**全部工作表**逐行抽取。
-/// 旧实现走 LibreOffice CSV 只导「活动工作表」，多 sheet 文件丢 90% 内容（实测 4 sheet
-/// 散热报告只抽到首页 418 字、CPU/温升数据全丢）。calamine 失败时（私有 .et 格式等）
-/// 回退 LibreOffice CSV，至少拿到首个工作表，不退化于旧行为。
-fn ingest_spreadsheet(
-    path: &Path,
-    basename: String,
-    path_str: String,
-    byte_size: u64,
-    kind: &str,
-) -> IngestResult {
-    // (markdown, warning)：calamine 成功取全表；失败回退 LibreOffice CSV（仅首个工作表）。
-    let (markdown, warning) = match spreadsheet_all_sheets_text(path) {
-        Ok(body) if !body.trim().is_empty() => (Some(body), None),
-        other => {
-            let fb_err = other.err();
-            match libreoffice_convert_text(path, "csv:Text - txt - csv (StarCalc)", "csv") {
-                Ok(content) => (
-                    Some(format!(
-                        "> 注：表格解析回退到 CSV，若原文件含多个工作表，此处仅含首个工作表。\n\n{content}"
-                    )),
-                    None,
-                ),
-                Err(e) => (None, Some(fb_err.unwrap_or(e))),
-            }
-        }
-    };
-    let token_estimate = markdown.as_deref().map(estimate_tokens).unwrap_or(0);
-    IngestResult {
-        kind: kind.into(),
-        basename,
-        path: path_str,
-        markdown,
-        token_estimate,
-        byte_size,
-        warning,
-    }
-}
-
-/// calamine 读电子表格全部工作表 → 逐行文本。每个工作表加 `## 工作表：名` 小标题，
-/// 每行用 ` | ` 连接非空尾部单元格（去掉行尾连续空单元格），整行空则跳过。
-/// 单元格内换行折成空格，避免破坏「一行 = 一条记录」的语义（利于切块/检索）。
-fn spreadsheet_all_sheets_text(path: &Path) -> Result<String, String> {
-    let bytes = std::fs::read(path).map_err(|e| format!("读取文件失败: {e}"))?;
-    spreadsheet_text_from_bytes(bytes)
-}
-
-/// 抽取核心（接收字节，便于单测喂 fixture）。见 [`spreadsheet_all_sheets_text`]。
-fn spreadsheet_text_from_bytes(bytes: Vec<u8>) -> Result<String, String> {
-    // DataType trait 提供 Data::is_empty()（calamine 0.26 是 trait 方法，非固有）。
-    use calamine::{open_workbook_auto_from_rs, Data, DataType, Reader};
-
-    let mut wb = open_workbook_auto_from_rs(std::io::Cursor::new(bytes))
-        .map_err(|e| format!("calamine 解析失败: {e}"))?;
-
-    let cell = |c: &Data| -> String {
-        if c.is_empty() {
-            String::new()
-        } else {
-            c.to_string()
-                .replace(['\n', '\r', '\t'], " ")
-                .trim()
-                .to_string()
-        }
-    };
-
-    let mut out = String::new();
-    let names: Vec<String> = wb.sheet_names().to_owned();
-    for name in names {
-        let range = match wb.worksheet_range(&name) {
-            Ok(r) => r,
-            Err(_) => continue,
-        };
-        if range.is_empty() {
-            continue;
-        }
-        out.push_str("## 工作表：");
-        out.push_str(&name);
-        out.push('\n');
-        for row in range.rows() {
-            let cells: Vec<String> = row.iter().map(&cell).collect();
-            let last = match cells.iter().rposition(|c| !c.is_empty()) {
-                Some(i) => i,
-                None => continue, // 整行空
-            };
-            out.push_str(&cells[..=last].join(" | "));
-            out.push('\n');
-        }
-        out.push('\n');
-    }
-    Ok(out)
-}
-
-/// 演示类（.pptx/.ppt/.odp + WPS .dps）：LibreOffice **没有 Impress→txt 导出**
-/// （实测无产出），所以先转 PDF 再用 pdftotext 抽每页文字（标题/要点/中文都保留）。
-/// 自己转出的 PDF 必有文字层，不需要扫描件 OCR 兜底。
-fn ingest_presentation(path: &Path, basename: String, byte_size: u64, kind: &str) -> IngestResult {
-    let tools = system_tools();
-    if !tools.libreoffice || !tools.pdftotext {
-        return IngestResult::warning(
-            kind,
-            &basename,
-            path,
-            byte_size,
-            crate::platform::os::presentation_pdf_missing_message(),
-        );
-    }
-    match libreoffice_presentation_text(path) {
-        Ok(content) if !content.trim().is_empty() => {
-            IngestResult::with_markdown(kind, &basename, path, byte_size, content)
-        }
-        Ok(_) => IngestResult::warning(kind, &basename, path, byte_size, "演示文稿未提取到文字"),
-        Err(e) => IngestResult::warning(kind, &basename, path, byte_size, e),
-    }
-}
-
-/// 演示文稿 → PDF（LibreOffice）→ pdftotext 的串联，返回纯文本。
-fn libreoffice_presentation_text(path: &Path) -> Result<String, String> {
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-pptpdf-{ts}"));
-    std::fs::create_dir_all(&tmpdir).map_err(|e| format!("创建临时目录失败: {e}"))?;
-
-    let convert = libreoffice_tool_command()
-        .arg(libreoffice_user_installation_arg(&tmpdir.join("profile"))?)
-        .arg("--headless")
-        .arg("--convert-to")
-        .arg("pdf")
-        .arg("--outdir")
-        .arg(&tmpdir)
-        .arg(path)
-        .output();
-
-    let result = match convert {
-        Ok(o) if o.status.success() => {
-            let stem = path
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .unwrap_or("converted");
-            let pdf_path = tmpdir.join(format!("{stem}.pdf"));
-            pdf_tool_command("pdftotext")
-                .arg("-layout")
-                .arg(&pdf_path)
-                .arg("-")
-                .output()
-                .map_err(|e| format!("pdftotext 调用失败: {e}"))
-                .and_then(|o| {
-                    if o.status.success() {
-                        Ok(String::from_utf8_lossy(&o.stdout).into_owned())
-                    } else {
-                        Err(format!(
-                            "pdftotext 失败: {}",
-                            String::from_utf8_lossy(&o.stderr).trim()
-                        ))
-                    }
-                })
-        }
-        Ok(o) => Err(format!(
-            "LibreOffice 转 PDF 失败: {}",
-            String::from_utf8_lossy(&o.stderr).trim()
-        )),
-        Err(e) => Err(format!("LibreOffice 调用失败: {e}")),
-    };
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-/// 图片：Qwen3.6 有视觉能力(2026-05-28 实证),不再跑 OCR 降级。这里只登记元
-/// 数据,真正"看图"由 LLM 在对话里调 `image_analyze` 完成——commands.rs 发消息时
-/// 会把图拷进 session workspace 的 `attachments/` 并给出相对路径引导。
-/// markdown 留空(不预解析像素),token_estimate=0(视觉 token 量取决于分辨率,
-/// 不在此处解码估算,UI 计数会略低,属已知局限)。
-fn ingest_image(_path: &Path, basename: String, path_str: String, byte_size: u64) -> IngestResult {
-    IngestResult::placeholder("image", &basename, Path::new(&path_str), byte_size)
-}
-
-/// 对单张图片跑 tesseract，识别文字到 stdout。`tesseract <img> - -l <langs>`。
-fn ocr_image(path: &Path) -> Result<String, String> {
-    let lang = ocr_lang_arg();
-    let mut command = ocr_tool_command();
-    command.arg(path).arg("-").arg("-l").arg(&lang);
-    add_ocr_tessdata_arg(&mut command);
-    let out = command
-        .output()
-        .map_err(|e| format!("tesseract 调用失败: {e}"))?;
-    if out.status.success() {
-        Ok(String::from_utf8_lossy(&out.stdout).trim_end().to_string())
-    } else {
-        Err(format!(
-            "tesseract 退出码 {:?}: {}",
-            out.status.code(),
-            String::from_utf8_lossy(&out.stderr).trim()
-        ))
-    }
-}
-
-/// 扫描件 PDF OCR 兜底：pdftoppm 逐页转 PNG（150 dpi），每页跑 tesseract，拼接。
-/// 页数封顶 PDF_OCR_MAX_PAGES，超出截断并在末尾标注，避免几十页扫描件把上下文撑爆。
-/// 知识库专用：对图片做 OCR 取文字。**只在 KB 入库调**——对话附件图仍走视觉(image_analyze)，
-/// `ingest_image` 不在这里 OCR（保留 2026-05-28「图片不预解析、交给视觉」的对话侧设计）。
-/// 没装 tesseract / OCR 失败 / 识别为空 → None（调用方落 skipped）。
-pub fn ocr_image_for_kb(path: &Path) -> Option<String> {
-    if !system_tools().tesseract {
-        return None;
-    }
-    match ocr_image(path) {
-        Ok(t) if !t.trim().is_empty() => Some(t),
-        _ => None,
-    }
-}
-
-fn ocr_pdf(path: &Path, basename: String, path_str: String, byte_size: u64) -> IngestResult {
-    const PDF_OCR_MAX_PAGES: u32 = 30;
-    let tools = system_tools();
-    // 构造器从 &Path 还原 path 字符串；path_str 来自上游 path.to_string_lossy()，
-    // 用 Path::new(&path_str) 复用同一字符串视图，保证 path 字段逐字节一致。
-    let result_path = Path::new(&path_str);
-    if !tools.tesseract || !tools.pdftoppm {
-        return IngestResult::warning(
-            "pdf",
-            &basename,
-            result_path,
-            byte_size,
-            crate::platform::os::pdf_ocr_missing_message(),
-        );
-    }
-
-    // 临时目录：每次唯一，避免并发冲突。
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-pdfocr-{ts}"));
-    if let Err(e) = std::fs::create_dir_all(&tmpdir) {
-        return IngestResult::warning(
-            "pdf",
-            &basename,
-            result_path,
-            byte_size,
-            format!("创建临时目录失败: {e}"),
-        );
-    }
-
-    let prefix = tmpdir.join("page");
-    // pdftoppm -png -r 150 -l <max> <pdf> <prefix> → prefix-1.png, prefix-2.png ...
-    let convert = pdf_tool_command("pdftoppm")
-        .arg("-png")
-        .arg("-r")
-        .arg("150")
-        .arg("-l")
-        .arg(PDF_OCR_MAX_PAGES.to_string())
-        .arg(path)
-        .arg(&prefix)
-        .output();
-
-    let result = match convert {
-        Ok(o) if o.status.success() => {
-            // 收集生成的 png，按文件名排序保证页序。
-            let mut pages: Vec<PathBuf> = std::fs::read_dir(&tmpdir)
-                .map(|rd| {
-                    rd.filter_map(|e| e.ok().map(|e| e.path()))
-                        .filter(|p| p.extension().and_then(|e| e.to_str()) == Some("png"))
-                        .collect()
-                })
-                .unwrap_or_default();
-            pages.sort();
-
-            if pages.is_empty() {
-                IngestResult::warning(
-                    "pdf",
-                    &basename,
-                    result_path,
-                    byte_size,
-                    "PDF 无文字层，且 pdftoppm 未产出可识别页",
-                )
-            } else {
-                let mut parts = Vec::new();
-                for (idx, page) in pages.iter().enumerate() {
-                    match ocr_image(page) {
-                        Ok(text) if !text.trim().is_empty() => {
-                            parts.push(format!("## 第 {} 页\n\n{}", idx + 1, text.trim()));
-                        }
-                        _ => {}
-                    }
-                }
-                let mut content = parts.join("\n\n");
-                if pages.len() as u32 >= PDF_OCR_MAX_PAGES {
-                    content.push_str(&format!(
-                        "\n\n> ⚠️ 扫描件页数较多，OCR 仅处理前 {PDF_OCR_MAX_PAGES} 页"
-                    ));
-                }
-                if content.trim().is_empty() {
-                    IngestResult::warning(
-                        "pdf",
-                        &basename,
-                        result_path,
-                        byte_size,
-                        "扫描件 OCR 未识别到文字",
-                    )
-                } else {
-                    // 同时带 markdown 与 warning（OCR 误差提示），不符合任何构造器，
-                    // 保留字面量 —— 这是「正文 + 告警」的特例。
-                    let tokens = estimate_tokens(&content);
-                    IngestResult {
-                        kind: "pdf".into(),
-                        basename,
-                        path: path_str.clone(),
-                        markdown: Some(content),
-                        token_estimate: tokens,
-                        byte_size,
-                        warning: Some("扫描件 PDF，内容由 OCR 提取，可能有识别误差".into()),
-                    }
-                }
-            }
-        }
-        Ok(o) => IngestResult::warning(
-            "pdf",
-            &basename,
-            result_path,
-            byte_size,
-            format!(
-                "pdftoppm 转图失败: {}",
-                String::from_utf8_lossy(&o.stderr).trim()
-            ),
-        ),
-        Err(e) => IngestResult::warning(
-            "pdf",
-            &basename,
-            result_path,
-            byte_size,
-            format!("pdftoppm 调用失败: {e}"),
-        ),
-    };
-
-    let _ = std::fs::remove_dir_all(&tmpdir);
-    result
-}
-
-/// 压缩包（.zip/.rar/.7z）：先用 7z 列出内容做炸弹预检（条目数 + 解压后总大小，
-/// 解压前就拦），通过后解压到临时目录，递归调主 `ingest` 处理每个文件并汇总。
-/// 嵌套压缩包不再展开（防套娃炸弹）。因为复用主 ingest，包里的 PDF/Office/图片
-/// 都会按各自管线（含 OCR）处理。
-fn ingest_archive(path: &Path, basename: String, path_str: String, byte_size: u64) -> IngestResult {
-    const MAX_ENTRIES: usize = 50;
-    const MAX_TOTAL_BYTES: u64 = 100 * 1024 * 1024; // 解压后总量上限，防压缩炸弹
-
-    let mk_err = |msg: String| {
-        IngestResult::warning("archive", &basename, Path::new(&path_str), byte_size, msg)
-    };
-
-    if !system_tools().sevenzip {
-        return mk_err(archive_tool_missing_message());
-    }
-
-    // 预检：解压前就用 7z 列表拦截压缩炸弹。
-    match archive_list_stats(path) {
-        Ok((count, total)) => {
-            if count > MAX_ENTRIES {
-                return mk_err(format!(
-                    "压缩包条目过多（{count} > {MAX_ENTRIES}），拒绝展开"
-                ));
-            }
-            if total > MAX_TOTAL_BYTES {
-                return mk_err(format!(
-                    "压缩包解压后约 {:.0} MB，超过 {} MB 上限（疑似压缩炸弹），拒绝展开",
-                    total as f64 / 1024.0 / 1024.0,
-                    MAX_TOTAL_BYTES / 1024 / 1024
-                ));
-            }
-        }
-        Err(e) => return mk_err(format!("压缩包内容读取失败: {e}")),
-    }
-
-    let ts = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    let tmpdir = std::env::temp_dir().join(format!("pinvou3-archive-{ts}"));
-    if let Err(e) = std::fs::create_dir_all(&tmpdir) {
-        return mk_err(format!("创建临时目录失败: {e}"));
-    }
-
-    let extract = archive_tool_command()
-        .arg("x")
-        .arg("-y")
-        .arg(format!("-o{}", tmpdir.display()))
-        .arg(path)
-        .output();
-    if !matches!(&extract, Ok(o) if o.status.success()) {
-        let detail = match extract {
-            Ok(o) => String::from_utf8_lossy(&o.stderr).trim().to_string(),
-            Err(e) => e.to_string(),
-        };
-        let _ = std::fs::remove_dir_all(&tmpdir);
-        return mk_err(format!("7z 解压失败: {detail}"));
-    }
-
-    // 递归收集文件，对每个调主 ingest；嵌套压缩包不展开。
-    let mut files = Vec::new();
-    collect_files(&tmpdir, &mut files, MAX_ENTRIES);
-
-    let mut sections = Vec::new();
-    for f in &files {
-        let rel = f
-            .strip_prefix(&tmpdir)
-            .unwrap_or(f)
-            .to_string_lossy()
-            .to_string();
-        let ext = f
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(|s| s.to_ascii_lowercase())
-            .unwrap_or_default();
-        if classify(&ext) == "archive" {
-            sections.push(format!("### {rel}\n⚠️ 嵌套压缩包，未展开（防套娃）\n"));
-            continue;
-        }
-        let r = ingest(f);
-        let body = match (&r.markdown, &r.warning) {
-            (Some(md), _) => md.clone(),
-            (None, Some(w)) => format!("⚠️ {w}"),
-            (None, None) => "(无文本内容)".to_string(),
-        };
-        sections.push(format!("### {rel} ({})\n{body}\n", r.kind));
-    }
-    let _ = std::fs::remove_dir_all(&tmpdir);
-
-    if sections.is_empty() {
-        return mk_err("压缩包为空或无可识别文件".into());
-    }
-    let content = format!(
-        "压缩包 {} 含 {} 个文件：\n\n{}",
-        basename,
-        files.len(),
-        sections.join("\n")
-    );
-    IngestResult::with_markdown("archive", &basename, path, byte_size, content)
-}
-
-/// `7z l -slt` 列出条目，返回 (文件数, 解压后总字节)。用于解压前的炸弹预检。
-fn archive_list_stats(path: &Path) -> Result<(usize, u64), String> {
-    let out = archive_tool_command()
-        .arg("l")
-        .arg("-slt")
-        .arg(path)
-        .output()
-        .map_err(|e| format!("7z 调用失败: {e}"))?;
-    if !out.status.success() {
-        return Err(String::from_utf8_lossy(&out.stderr).trim().to_string());
-    }
-    let text = String::from_utf8_lossy(&out.stdout);
-    let total: u64 = text
-        .lines()
-        .filter_map(|l| l.trim().strip_prefix("Size = "))
-        .filter_map(|v| v.trim().parse::<u64>().ok())
-        .sum();
-    // `-slt` 的第一个 "Path =" 块是归档自身，扣掉。
-    let paths = text
-        .lines()
-        .filter(|l| l.trim_start().starts_with("Path = "))
-        .count();
-    Ok((paths.saturating_sub(1), total))
-}
-
-fn archive_tool_missing_message() -> String {
-    if crate::platform::os::show_archive_dependency_check() {
-        let packages = crate::platform::os::archive_dependency_packages();
-        if packages.trim().is_empty() {
-            "压缩包解析需要 7z，请按当前系统方式安装压缩包解析工具".into()
-        } else {
-            format!("压缩包解析需要 7z: sudo apt install {packages}")
-        }
-    } else {
-        "内置压缩包解析组件缺失或不可用，请修复或重新安装 pinvou。".into()
-    }
-}
-
-/// 递归收集目录下的普通文件（不含目录本身），到达 `limit` 即停。
-fn collect_files(dir: &Path, out: &mut Vec<PathBuf>, limit: usize) {
-    if out.len() >= limit {
-        return;
-    }
-    let Ok(rd) = std::fs::read_dir(dir) else {
-        return;
-    };
-    for entry in rd.flatten() {
-        if out.len() >= limit {
-            return;
-        }
-        let p = entry.path();
-        if p.is_dir() {
-            collect_files(&p, out, limit);
-        } else if p.is_file() {
-            out.push(p);
-        }
-    }
-}
-
-/// 邮件（.eml / .msg）：.eml 直接用 python 标准库 email 模块解出收发件人/主题/
-/// 日期/正文/附件名；.msg 在 Windows 走 Rust 原生解析，非 Windows 保留 msgconvert 转 .eml。
-fn ingest_email(
-    path: &Path,
-    basename: String,
-    path_str: String,
-    byte_size: u64,
-    kind: &str,
-) -> IngestResult {
-    let tools = system_tools();
-    let mk = |markdown: Option<String>, warning: Option<String>| {
-        let token_estimate = markdown.as_deref().map(estimate_tokens).unwrap_or(0);
-        IngestResult {
-            kind: kind.into(),
-            basename: basename.clone(),
-            path: path_str.clone(),
-            markdown,
-            token_estimate,
-            byte_size,
-            warning,
-        }
-    };
-
-    if kind == "msg" && crate::platform::os::msg_native_supported() {
-        return match parse_msg_via_msg_parser(path) {
-            Ok(text) => mk(Some(text), None),
-            Err(e) => mk(None, Some(e)),
-        };
-    }
-
-    if !tools.python {
-        return mk(None, Some("邮件解析需要可用的 Python 运行时。".into()));
-    }
-
-    let parsed = if kind == "msg" {
-        if !tools.msgconvert {
-            return mk(
-                None,
-                Some(".msg 解析需要: sudo apt install libemail-outlook-message-perl".into()),
-            );
-        }
-        // msgconvert 把 .msg 转成 .eml（输出到 cwd），用临时目录承接再解析。
-        let ts = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_nanos())
-            .unwrap_or(0);
-        let tmpdir = std::env::temp_dir().join(format!("pinvou3-msg-{ts}"));
-        if let Err(e) = std::fs::create_dir_all(&tmpdir) {
-            return mk(None, Some(format!("创建临时目录失败: {e}")));
-        }
-        let conv = Command::new("msgconvert")
-            .current_dir(&tmpdir)
-            .arg(path)
-            .output();
-        let result = if matches!(&conv, Ok(o) if o.status.success()) {
-            let stem = path.file_stem().and_then(|s| s.to_str()).unwrap_or("mail");
-            let eml = tmpdir.join(format!("{stem}.eml"));
-            parse_eml_via_python(&eml)
-        } else {
-            let detail = match conv {
-                Ok(o) => String::from_utf8_lossy(&o.stderr).trim().to_string(),
-                Err(e) => e.to_string(),
-            };
-            Err(format!("msgconvert 转换失败: {detail}"))
-        };
-        let _ = std::fs::remove_dir_all(&tmpdir);
-        result
-    } else {
-        parse_eml_via_python(path)
-    };
-
-    match parsed {
-        Ok(text) => mk(Some(text), None),
-        Err(e) => mk(None, Some(e)),
-    }
-}
-
-/// Outlook .msg 解析结果格式化为与 .eml 接近的可读邮件文本。
-#[derive(Default)]
-struct MsgMarkdownParts {
-    sender: String,
-    to: Vec<String>,
-    cc: Vec<String>,
-    bcc: Vec<String>,
-    subject: String,
-    date: String,
-    body: String,
-    attachments: Vec<String>,
-}
-
-fn parse_msg_via_msg_parser(path: &Path) -> Result<String, String> {
-    let outlook =
-        msg_parser::Outlook::from_path(path).map_err(|e| format!(".msg 解析失败: {e}"))?;
-    let body = decode_msg_body(&outlook);
-    let parts = MsgMarkdownParts {
-        sender: person_to_text(&outlook.sender),
-        to: people_to_text(&outlook.to),
-        cc: people_to_text(&outlook.cc),
-        bcc: people_to_text(&outlook.bcc),
-        subject: clean_msg_text(&outlook.subject),
-        date: first_non_empty([
-            outlook.message_delivery_time.as_str(),
-            outlook.client_submit_time.as_str(),
-            outlook.headers.date.as_str(),
-            outlook.creation_time.as_str(),
-        ]),
-        body,
-        attachments: outlook
-            .attachments
-            .iter()
-            .map(attachment_to_name)
-            .filter(|name| !name.is_empty())
-            .collect(),
-    };
-    let markdown = format_msg_as_markdown(&parts);
-    if markdown.trim().is_empty() {
-        Err(".msg 解析失败: 未提取到邮件内容".into())
-    } else {
-        Ok(markdown)
-    }
-}
-
-fn format_msg_as_markdown(parts: &MsgMarkdownParts) -> String {
-    let mut out = String::new();
-    push_mail_line(&mut out, "发件人", &parts.sender);
-    push_mail_line(&mut out, "收件人", &parts.to.join(", "));
-    push_mail_line(&mut out, "抄送", &parts.cc.join(", "));
-    push_mail_line(&mut out, "密送", &parts.bcc.join(", "));
-    push_mail_line(&mut out, "主题", &parts.subject);
-    push_mail_line(&mut out, "日期", &parts.date);
-    if !parts.body.trim().is_empty() {
-        if !out.is_empty() {
-            out.push('\n');
-        }
-        out.push_str("正文:\n");
-        out.push_str(parts.body.trim());
-        out.push('\n');
-    }
-    if !parts.attachments.is_empty() {
-        if !out.is_empty() {
-            out.push('\n');
-        }
-        out.push_str("附件: ");
-        out.push_str(&parts.attachments.join(", "));
-    }
-    out.trim_end().to_string()
-}
-
-fn push_mail_line(out: &mut String, label: &str, value: &str) {
-    let value = value.trim();
-    if value.is_empty() {
-        return;
-    }
-    out.push_str(label);
-    out.push_str(": ");
-    out.push_str(value);
-    out.push('\n');
-}
-
-fn people_to_text(people: &[msg_parser::Person]) -> Vec<String> {
-    people
-        .iter()
-        .map(person_to_text)
-        .filter(|value| !value.is_empty())
-        .collect()
-}
-
-fn person_to_text(person: &msg_parser::Person) -> String {
-    clean_msg_text(&person.to_string())
-}
-
-fn attachment_to_name(attachment: &msg_parser::Attachment) -> String {
-    [
-        &attachment.long_file_name,
-        &attachment.file_name,
-        &attachment.display_name,
-    ]
-    .into_iter()
-    .map(|value| clean_msg_text(value))
-    .find(|value| !value.is_empty())
-    .unwrap_or_default()
-}
-
-fn first_non_empty<'a>(values: impl IntoIterator<Item = &'a str>) -> String {
-    values
-        .into_iter()
-        .map(clean_msg_text)
-        .find(|value| !value.is_empty())
-        .unwrap_or_default()
-}
-
-fn decode_msg_body(outlook: &msg_parser::Outlook) -> String {
-    let body = clean_msg_text(&outlook.body);
-    if !body.is_empty() {
-        return body;
-    }
-
-    let html = if outlook.html.trim().is_empty() {
-        outlook.html_from_rtf().unwrap_or_default()
-    } else {
-        outlook.html.clone()
-    };
-    let decoded = decode_msg_html_payload(&html);
-    let text = html_to_text(&decoded);
-    if text.is_empty() {
-        decoded
-    } else {
-        text
-    }
-}
-
-fn clean_msg_text(value: &str) -> String {
-    value
-        .chars()
-        .filter(|ch| *ch != '\0')
-        .collect::<String>()
-        .trim()
-        .to_string()
-}
-
-fn decode_msg_html_payload(value: &str) -> String {
-    let value = clean_msg_text(value);
-    if value.len() < 8
-        || !value.len().is_multiple_of(2)
-        || !value.bytes().all(|b| b.is_ascii_hexdigit())
-    {
-        return value;
-    }
-    let mut bytes = Vec::with_capacity(value.len() / 2);
-    let raw = value.as_bytes();
-    for pair in raw.chunks_exact(2) {
-        let Ok(hex) = std::str::from_utf8(pair) else {
-            return value;
-        };
-        let Ok(byte) = u8::from_str_radix(hex, 16) else {
-            return value;
-        };
-        bytes.push(byte);
-    }
-    decode_msg_bytes(&bytes)
-}
-
-fn decode_msg_bytes(bytes: &[u8]) -> String {
-    if bytes.starts_with(&[0xFF, 0xFE]) {
-        return decode_utf16le(&bytes[2..]);
-    }
-    if bytes.starts_with(&[0xFE, 0xFF]) {
-        return decode_utf16be(&bytes[2..]);
-    }
-    if let Ok(text) = String::from_utf8(bytes.to_vec()) {
-        return clean_msg_text(&text);
-    }
-    let nul_count = bytes.iter().filter(|byte| **byte == 0).count();
-    if nul_count > bytes.len() / 4 {
-        decode_utf16le(bytes)
-    } else {
-        clean_msg_text(&String::from_utf8_lossy(bytes))
-    }
-}
-
-fn decode_utf16le(bytes: &[u8]) -> String {
-    let units: Vec<u16> = bytes
-        .chunks_exact(2)
-        .map(|chunk| u16::from_le_bytes([chunk[0], chunk[1]]))
-        .collect();
-    clean_msg_text(&String::from_utf16_lossy(&units))
-}
-
-fn decode_utf16be(bytes: &[u8]) -> String {
-    let units: Vec<u16> = bytes
-        .chunks_exact(2)
-        .map(|chunk| u16::from_be_bytes([chunk[0], chunk[1]]))
-        .collect();
-    clean_msg_text(&String::from_utf16_lossy(&units))
-}
-
-fn html_to_text(html: &str) -> String {
-    let html = remove_html_section(html, "script");
-    let html = remove_html_section(&html, "style");
-    let html = remove_html_section(&html, "head");
-    let html = html
-        .replace("<br>", "\n")
-        .replace("<br/>", "\n")
-        .replace("<br />", "\n")
-        .replace("</p>", "\n")
-        .replace("</div>", "\n")
-        .replace("</tr>", "\n")
-        .replace("</li>", "\n");
-
-    let mut out = String::with_capacity(html.len());
-    let mut in_tag = false;
-    for ch in html.chars() {
-        match ch {
-            '<' => in_tag = true,
-            '>' => {
-                in_tag = false;
-                out.push(' ');
-            }
-            _ if !in_tag => out.push(ch),
-            _ => {}
-        }
-    }
-    collapse_text(&decode_html_entities(&out))
-}
-
-fn remove_html_section(input: &str, tag: &str) -> String {
-    let mut out = input.to_string();
-    let open = format!("<{tag}");
-    let close = format!("</{tag}>");
-    loop {
-        let lower = out.to_ascii_lowercase();
-        let Some(start) = lower.find(&open) else {
-            break;
-        };
-        let Some(end_rel) = lower[start..].find(&close) else {
-            out.truncate(start);
-            break;
-        };
-        let end = start + end_rel + close.len();
-        out.replace_range(start..end, " ");
-    }
-    out
-}
-
-fn decode_html_entities(value: &str) -> String {
-    let mut out = String::with_capacity(value.len());
-    let mut rest = value;
-    while let Some(start) = rest.find('&') {
-        out.push_str(&rest[..start]);
-        let after_amp = &rest[start + 1..];
-        let Some(end) = after_amp.find(';') else {
-            out.push('&');
-            rest = after_amp;
-            continue;
-        };
-        let entity = &after_amp[..end];
-        if let Some(decoded) = decode_html_entity(entity) {
-            out.push(decoded);
-        } else {
-            out.push('&');
-            out.push_str(entity);
-            out.push(';');
-        }
-        rest = &after_amp[end + 1..];
-    }
-    out.push_str(rest);
-    out
-}
-
-fn decode_html_entity(entity: &str) -> Option<char> {
-    match entity {
-        "amp" => Some('&'),
-        "lt" => Some('<'),
-        "gt" => Some('>'),
-        "quot" => Some('"'),
-        "apos" => Some('\''),
-        "nbsp" => Some(' '),
-        _ if entity.starts_with("#x") || entity.starts_with("#X") => {
-            u32::from_str_radix(&entity[2..], 16)
-                .ok()
-                .and_then(char::from_u32)
-        }
-        _ if entity.starts_with('#') => entity[1..].parse::<u32>().ok().and_then(char::from_u32),
-        _ => None,
-    }
-}
-
-fn collapse_text(value: &str) -> String {
-    let mut out = String::new();
-    let mut blank_lines = 0;
-    for line in value.lines() {
-        let line = line.split_whitespace().collect::<Vec<_>>().join(" ");
-        if line.is_empty() {
-            blank_lines += 1;
-            if blank_lines <= 1 && !out.is_empty() {
-                out.push('\n');
-            }
-        } else {
-            blank_lines = 0;
-            if !out.is_empty() && !out.ends_with('\n') {
-                out.push('\n');
-            }
-            out.push_str(&line);
-        }
-    }
-    out.trim().to_string()
-}
-
-/// 用 python 标准库 email 模块把 .eml 解析成可读文本（收发件人/主题/日期/正文/
-/// 附件名）。脚本走 stdin 之外的 argv[1] 取路径，正文优先纯文本、回退 HTML。
-fn parse_eml_via_python(path: &Path) -> Result<String, String> {
-    const SCRIPT: &str = r#"
-import sys, email
-from email import policy
-with open(sys.argv[1], 'rb') as f:
-    msg = email.message_from_binary_file(f, policy=policy.default)
-def h(k):
-    v = msg[k]
-    return str(v) if v else ''
-print('发件人:', h('from'))
-print('收件人:', h('to'))
-if msg['cc']:
-    print('抄送:', h('cc'))
-print('主题:', h('subject'))
-print('日期:', h('date'))
-try:
-    body = msg.get_body(preferencelist=('plain', 'html'))
-    if body is not None:
-        print('\n正文:')
-        print(body.get_content())
-except Exception as e:
-    print('\n(正文解析失败:', e, ')')
-atts = [p.get_filename() for p in msg.iter_attachments() if p.get_filename()]
-if atts:
-    print('\n附件:', ', '.join(atts))
-"#;
-    let program = crate::platform::paths::python_command();
-    let out = crate::platform::process::HiddenCommand::new(&program)
-        .arg("-c")
-        .arg(SCRIPT)
-        .arg(path)
-        .output()
-        .map_err(|e| format!("Python 调用失败({program}): {e}"))?;
-    if out.status.success() {
-        Ok(String::from_utf8_lossy(&out.stdout).trim_end().to_string())
-    } else {
-        Err(format!(
-            "邮件解析失败: {}",
-            String::from_utf8_lossy(&out.stderr).trim()
-        ))
-    }
-}
-
 /// 音视频：本地语音转录（whisper 等）尚未部署到 GB10，先优雅降级，明确告知用户
 /// 「未处理」而非臆测内容。真正转录留作未来独立能力（见 process.md）。
 fn media_placeholder(basename: String, path_str: String, byte_size: u64) -> IngestResult {
@@ -1898,230 +301,6 @@ fn secret_placeholder(basename: String, path_str: String, byte_size: u64) -> Ing
         byte_size,
         "检测到密钥/私钥文件,为防止泄露给 LLM 已拒绝读取内容",
     )
-}
-
-/// 内容侧私钥检测:捕获改了名 / 无扩展名 / 套了 .txt 外壳的私钥(如 `id_rsa`、
-/// `server-key`、`backup.txt` 里贴了 PEM)。PEM 私钥头和 OpenSSH magic 都在文件开头,
-/// 故只看前 8 KiB 即可。**不**匹配 `-----BEGIN CERTIFICATE-----`(证书可公开)。
-fn looks_like_secret_material(bytes: &[u8]) -> bool {
-    const MARKERS: &[&[u8]] = &[
-        b"-----BEGIN PRIVATE KEY-----",
-        b"-----BEGIN RSA PRIVATE KEY-----",
-        b"-----BEGIN DSA PRIVATE KEY-----",
-        b"-----BEGIN EC PRIVATE KEY-----",
-        b"-----BEGIN OPENSSH PRIVATE KEY-----",
-        b"-----BEGIN ENCRYPTED PRIVATE KEY-----",
-        b"-----BEGIN PGP PRIVATE KEY BLOCK-----",
-        b"openssh-key-v1",
-    ];
-    let head = &bytes[..bytes.len().min(8192)];
-    MARKERS
-        .iter()
-        .any(|m| head.windows(m.len()).any(|w| w == *m))
-}
-
-/// 对字节内容做嗅探,判定是文本还是二进制。只看前 `SNIFF_BYTES` 字节即可下结论
-/// (NUL / 控制字符都在头部就能判定),**但调用方仍会 `std::fs::read` 整个文件** ——
-/// 因为判定为文本时还需要全部字节做解码。`ingest` 入口已用 MAX_FILE_BYTES(20MiB)
-/// 挡住超大文件,故整文件读取的内存上界就是 20MiB,安全。
-///
-/// 算法(业界标准,`file` / `grep -I` / vim 同源):
-/// 1. 含 NUL 字节(0x00)→ 二进制(确定性最强,文本文件几乎不会含 NUL);
-/// 2. 控制字符(除 `\t \n \r \f`)占比 > 30% → 二进制;
-/// 3. 否则 → 文本。
-fn looks_like_binary_sample(bytes: &[u8]) -> bool {
-    const SNIFF_BYTES: usize = 8192;
-    let sample = if bytes.len() > SNIFF_BYTES {
-        &bytes[..SNIFF_BYTES]
-    } else {
-        bytes
-    };
-    if sample.is_empty() {
-        return false; // 空文件不当二进制(此处的 sample.is_empty() 判定兜底)
-    }
-    // NUL 字节 → 二进制(强信号)。
-    if sample.contains(&0u8) {
-        return true;
-    }
-    // 控制字符占比:除 \t(0x09) \n(0x0A) \r(0x0D) \f(0x0C) 外的 0x01..=0x1F。
-    // > 30% 视为二进制(纯文本通常 < 5%,二进制常 > 50%)。
-    let ctrl = sample
-        .iter()
-        .filter(|&&b| b < 0x20 && !matches!(b, 0x09 | 0x0A | 0x0D | 0x0C))
-        .count();
-    ctrl * 10 > sample.len() * 3 // ctrl/sample > 0.3,用整数乘法避浮点
-}
-
-/// 未知扩展名 / 无扩展名文件的兜底:先嗅探,是文本就当文本读(让模型看到内容),
-/// 真二进制才回 binary_placeholder。
-///
-/// 非 UTF-8 文本优先按 GB18030(兼容 GBK)解码；只有 GB18030 也不合法时才 lossy。
-fn sniff_text_or_binary(
-    path: &Path,
-    basename: String,
-    path_str: String,
-    byte_size: u64,
-) -> IngestResult {
-    let bytes = match std::fs::read(path) {
-        Ok(b) => b,
-        // 读失败(权限/TOCTOU 等)如实上报,不要误报成「不支持的二进制类型」。
-        Err(e) => {
-            return IngestResult::warning(
-                "binary",
-                &basename,
-                path,
-                byte_size,
-                format!("文件读取失败: {e}"),
-            );
-        }
-    };
-    // 内容侧安全兜底:无扩展名 / 改名的私钥(id_rsa、server-key 等)在此拦截。
-    if looks_like_secret_material(&bytes) {
-        return secret_placeholder(basename, path_str, byte_size);
-    }
-    // UTF-16 无 BOM 的 ASCII/混合文本含大量 NUL，本来会被二进制嗅探拒绝；先识别
-    // 奇偶字节 NUL 分布的强信号，确认不是 UTF-16 后再执行二进制降级。
-    if detect_utf16(&bytes).is_none() && looks_like_binary_sample(&bytes) {
-        return binary_placeholder(basename, path_str, byte_size);
-    }
-    // 判定为文本:统一解码(UTF-8 / UTF-16 / GB18030 / 最终 lossy 兜底)。
-    let (content, encoding_warning) = decode_text_bytes(&bytes);
-    let tokens = content.as_ref().map(|c| estimate_tokens(c)).unwrap_or(0);
-    IngestResult {
-        kind: "text".into(),
-        basename,
-        path: path_str,
-        markdown: content,
-        token_estimate: tokens,
-        byte_size,
-        warning: encoding_warning,
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Utf16Endian {
-    Little,
-    Big,
-}
-
-/// 识别 UTF-16 编码并返回 `(字节序, BOM 长度)`。
-///
-/// 有 BOM 时是确定性判断；无 BOM 时只接受强信号：至少两个 code unit 的同一侧为 NUL、
-/// 覆盖至少一半采样对，且另一侧 NUL 数不到其四分之一。这能接住常见 ASCII/中文混合
-/// UTF-16，同时避免仅凭单个 NUL 把普通二进制误判成文本。
-fn detect_utf16(bytes: &[u8]) -> Option<(Utf16Endian, usize)> {
-    if bytes.starts_with(&[0xFF, 0xFE]) {
-        return Some((Utf16Endian::Little, 2));
-    }
-    if bytes.starts_with(&[0xFE, 0xFF]) {
-        return Some((Utf16Endian::Big, 2));
-    }
-    if bytes.len() < 4 || !bytes.len().is_multiple_of(2) {
-        return None;
-    }
-
-    let sample_len = bytes.len().min(8192) & !1;
-    let sample = &bytes[..sample_len];
-    let pairs = sample.len() / 2;
-    let even_nuls = sample.iter().step_by(2).filter(|&&b| b == 0).count();
-    let odd_nuls = sample
-        .iter()
-        .skip(1)
-        .step_by(2)
-        .filter(|&&b| b == 0)
-        .count();
-    let strong = |dominant: usize, other: usize| {
-        dominant >= 2 && dominant * 2 >= pairs && dominant > other.saturating_mul(4)
-    };
-
-    if strong(odd_nuls, even_nuls) {
-        Some((Utf16Endian::Little, 0))
-    } else if strong(even_nuls, odd_nuls) {
-        Some((Utf16Endian::Big, 0))
-    } else {
-        None
-    }
-}
-
-fn decode_utf16(
-    bytes: &[u8],
-    endian: Utf16Endian,
-    bom_len: usize,
-) -> (Option<String>, Option<String>) {
-    let payload = &bytes[bom_len..];
-    let label = match endian {
-        Utf16Endian::Little => "UTF-16 LE",
-        Utf16Endian::Big => "UTF-16 BE",
-    };
-    if !payload.len().is_multiple_of(2) {
-        return (
-            None,
-            Some(format!("检测到 {label} 编码,但文件字节数不完整,未读取内容")),
-        );
-    }
-    let units: Vec<u16> = payload
-        .chunks_exact(2)
-        .map(|pair| match endian {
-            Utf16Endian::Little => u16::from_le_bytes([pair[0], pair[1]]),
-            Utf16Endian::Big => u16::from_be_bytes([pair[0], pair[1]]),
-        })
-        .collect();
-    match String::from_utf16(&units) {
-        Ok(text) => {
-            let bom = if bom_len == 0 { "无 BOM " } else { "" };
-            (
-                Some(text.trim_start_matches('\u{feff}').to_string()),
-                Some(format!("检测到 {bom}{label} 编码,已转换为 UTF-8")),
-            )
-        }
-        Err(_) => (
-            None,
-            Some(format!("检测到 {label} 编码,但内容损坏,未读取内容")),
-        ),
-    }
-}
-
-/// 把已读到的字节统一解码成文本内容,供 `ingest_text` 与 `sniff_text_or_binary` 复用。
-/// 返回 `(内容, 警告)`:
-/// - **UTF-16 LE / BE**:BOM 确定识别；无 BOM 仅按强 NUL 分布信号识别并转 UTF-8；
-/// - **UTF-8 BOM**(`EF BB BF` → U+FEFF):去掉开头 BOM,与 `ingest_office_text`
-///   的 `s.trim_start_matches('\u{feff}')` 处理一致,避免污染正文开头。
-/// - **严格 UTF-8**:直接用,无 warning。
-/// - **GB18030 / GBK**:严格解码成 Unicode,标 warning 说明发生了编码转换；
-/// - **其他非 UTF-8**:最后才 `from_utf8_lossy`,并明确提示可能有替换符。
-fn decode_text_bytes(bytes: &[u8]) -> (Option<String>, Option<String>) {
-    if let Some((endian, bom_len)) = detect_utf16(bytes) {
-        return decode_utf16(bytes, endian, bom_len);
-    }
-    match std::str::from_utf8(bytes) {
-        Ok(s) => {
-            let s = s.strip_prefix('\u{feff}').unwrap_or(s);
-            (Some(s.to_string()), None)
-        }
-        Err(_) => {
-            if let Some(decoded) =
-                encoding_rs::GB18030.decode_without_bom_handling_and_without_replacement(bytes)
-            {
-                return (
-                    Some(decoded.into_owned()),
-                    Some("文件非 UTF-8 编码,已按 GB18030/GBK 转换为 UTF-8".into()),
-                );
-            }
-            let lossy = String::from_utf8_lossy(bytes);
-            let s = lossy.strip_prefix('\u{feff}').unwrap_or(&lossy);
-            (
-                Some(s.to_string()),
-                Some("文件非 UTF-8 编码,已尽力还原(部分字符可能为替换符)".into()),
-            )
-        }
-    }
-}
-
-/// 粗算 token：中英混合按 `chars / 1.6` —— 比较保守，偏向高估避免炸上下文。
-/// 实测 cl100k_base 中文 1.0-1.5 char/token、英文 3-4 char/token。
-fn estimate_tokens(text: &str) -> u32 {
-    let chars = text.chars().count();
-    (chars as f64 / 1.6).ceil() as u32
 }
 
 /// 把剪贴板粘贴的图片 bytes 写到 `~/.pinvou3/pastes/<timestamp>-<sanitized_name>`。
@@ -2223,6 +402,7 @@ fn normalize_validated_path(path: &Path) -> PathBuf {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::process::Command;
 
     #[test]
     fn ingest_result_constructors_match_equivalent_literals() {
@@ -2370,24 +550,6 @@ mod tests {
         let p = std::env::temp_dir().join(format!("pinvou3-sniff-{}-{name}", std::process::id()));
         std::fs::write(&p, bytes).unwrap();
         p
-    }
-
-    #[test]
-    fn looks_like_binary_detects_nul_and_control_chars() {
-        // NUL 字节 → 二进制(强信号)。
-        assert!(looks_like_binary_sample(b"abc\x00def"));
-        // 纯文本(含 UTF-8 多字节中文)→ 非二进制。
-        assert!(!looks_like_binary_sample(
-            "hello world\n中文测试".as_bytes()
-        ));
-        // 大量控制字符 → 二进制。
-        let mut ctrl = vec![0x01u8; 1000];
-        ctrl.extend_from_slice(b"text");
-        assert!(looks_like_binary_sample(&ctrl));
-        // 正常含 \t \n \r → 非二进制。
-        assert!(!looks_like_binary_sample(b"col1\tcol2\nrow1\r\nrow2"));
-        // 空文件 → 非二进制(上层 byte_size==0 已挡)。
-        assert!(!looks_like_binary_sample(b""));
     }
 
     #[test]
@@ -2600,32 +762,6 @@ mod tests {
     }
 
     #[test]
-    fn looks_like_binary_30_percent_boundary() {
-        // 控制字符占比阈值:严格 > 30% 才判二进制(ctrl*10 > len*3)。
-        // 30/100 = 30% → 非二进制(边界外);31/100 > 30% → 二进制。
-        let text_byte = b'x';
-        let ctrl_byte = 0x01u8;
-        let make = |ctrl: usize, total: usize| {
-            let mut v = vec![text_byte; total];
-            for b in v.iter_mut().take(ctrl) {
-                *b = ctrl_byte;
-            }
-            v
-        };
-        assert!(
-            !looks_like_binary_sample(&make(30, 100)),
-            "恰好 30% 控制字符应为文本(严格大于才判二进制)"
-        );
-        assert!(
-            looks_like_binary_sample(&make(31, 100)),
-            "31% 控制字符应为二进制"
-        );
-        // 8192 采样边界:8192 全采,8193 截断到 8192 —— 只要前 8192 判定一致即可。
-        let mut big_text = vec![text_byte; 8193];
-        big_text[0] = 0x00; // 头部一个 NUL → 二进制(采样必命中)
-        assert!(looks_like_binary_sample(&big_text));
-    }
-    #[test]
     fn ingest_text_reads_md() {
         // 写一个临时文件，调 ingest
         let tmp = std::env::temp_dir().join("pinvou3-ingest-test.md");
@@ -2635,33 +771,6 @@ mod tests {
         assert!(r.markdown.as_deref().unwrap_or("").contains("标题"));
         assert!(r.token_estimate > 0);
         std::fs::remove_file(&tmp).ok();
-    }
-
-    #[test]
-    fn spreadsheet_extracts_all_sheets() {
-        // 防回归：旧实现走 LibreOffice CSV 只导「活动工作表」，多 sheet 文件丢 90% 内容
-        // （实测 4-sheet 散热报告只抽到首页、CPU/温升数据全失）。calamine 必须把
-        // 全部工作表逐行抽出。fixture 是 3-sheet 合成表（Cover/配置表/温升表）。
-        let bytes = include_bytes!("../../../test-fixtures/multi_sheet.xlsx").to_vec();
-        let txt = spreadsheet_text_from_bytes(bytes).expect("calamine 应能解析 fixture");
-        // 三个工作表标题都要在
-        assert!(txt.contains("## 工作表：Cover"), "缺 Cover sheet");
-        assert!(
-            txt.contains("## 工作表：System configuration"),
-            "缺配置表 sheet"
-        );
-        assert!(
-            txt.contains("## 工作表：Thermal test result"),
-            "缺温升表 sheet"
-        );
-        // 非首表的关键内容（旧 CSV 实现会整段丢失）
-        assert!(txt.contains("Ultra 7 258V"), "应抽到第二表的 CPU 型号");
-        assert!(txt.contains("83.6"), "应抽到第三表的温度实测值");
-        // 行结构保留：同一行单元格用 ' | ' 连接
-        assert!(
-            txt.contains("CPU | key part | model | Ultra 7 258V SRPMN"),
-            "同一行单元格应保留在一行"
-        );
     }
 
     #[test]
@@ -2700,14 +809,6 @@ mod tests {
     }
 
     #[test]
-    fn libreoffice_user_installation_uses_encoded_file_url() {
-        let arg = libreoffice_user_installation_arg(&std::env::temp_dir().join("profile dir"))
-            .expect("file URL");
-        assert!(arg.starts_with("-env:UserInstallation=file://"));
-        assert!(arg.ends_with("profile%20dir"));
-    }
-
-    #[test]
     fn validate_path_accepts_windows_canonicalized_home_file() {
         if !crate::platform::capabilities::is_windows() {
             return;
@@ -2736,143 +837,6 @@ mod tests {
     }
 
     #[test]
-    fn pdf_tool_command_uses_os_layer_program() {
-        let command = pdf_tool_command("pdftotext");
-        assert_eq!(
-            command.get_program(),
-            crate::platform::os::pdf_tool_path("pdftotext").as_os_str()
-        );
-    }
-
-    #[test]
-    fn archive_tool_command_uses_os_layer_program() {
-        let command = archive_tool_command();
-        assert_eq!(
-            command.get_program(),
-            crate::platform::os::archive_tool_path().as_os_str()
-        );
-    }
-
-    #[test]
-    fn dependency_check_respects_pdf_visibility_policy() {
-        let deps = check_dependencies();
-        let has_pdf = deps.iter().any(|item| item.key == "pdf");
-        let has_pandoc = deps.iter().any(|item| item.key == "office_modern");
-        let has_ocr = deps.iter().any(|item| item.key == "ocr");
-        let has_archive = deps.iter().any(|item| item.key == "archive");
-        assert_eq!(has_pdf, crate::platform::os::show_pdf_dependency_check());
-        assert_eq!(
-            has_pandoc,
-            crate::platform::os::show_pandoc_dependency_check()
-        );
-        assert_eq!(has_ocr, crate::platform::os::show_ocr_dependency_check());
-        assert_eq!(
-            has_archive,
-            crate::platform::os::show_archive_dependency_check()
-        );
-
-        if !crate::platform::os::show_pdf_dependency_check() {
-            assert!(
-                deps.iter()
-                    .all(|item| !item.apt.contains("poppler") && !item.apt.contains("pdfto")),
-                "hidden Windows Poppler dependency should not leave install hints: {deps:?}"
-            );
-        }
-        if !crate::platform::os::show_pandoc_dependency_check() {
-            assert!(
-                deps.iter()
-                    .all(|item| !item.apt.contains("pandoc") && item.key != "office_modern"),
-                "hidden Windows Pandoc dependency should not leave install hints: {deps:?}"
-            );
-        }
-        if !crate::platform::os::show_ocr_dependency_check() {
-            assert!(
-                deps.iter().all(|item| {
-                    !item.apt.contains("tesseract") && !item.apt.contains("tesseract-ocr")
-                }),
-                "hidden Windows OCR dependency should not leave install hints: {deps:?}"
-            );
-        }
-        if !crate::platform::os::show_archive_dependency_check() {
-            assert!(
-                deps.iter()
-                    .all(|item| !item.apt.contains("p7zip") && item.key != "archive"),
-                "hidden Windows archive dependency should not leave install hints: {deps:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn windows_archive_missing_message_points_to_bundled_runtime() {
-        if !crate::platform::capabilities::is_windows() {
-            return;
-        }
-        let message = archive_tool_missing_message();
-
-        assert!(message.contains("内置压缩包解析组件"));
-        assert!(!message.contains("sudo apt install"));
-        assert!(!message.contains("p7zip-full"));
-    }
-
-    #[test]
-    fn msg_markdown_format_includes_headers_body_and_attachments() {
-        let parts = MsgMarkdownParts {
-            sender: "alice@example.com".into(),
-            to: vec!["bob@example.com".into()],
-            cc: vec!["carol@example.com".into()],
-            bcc: vec!["audit@example.com".into()],
-            subject: "项目进展".into(),
-            date: "2026-06-25T10:00:00Z".into(),
-            body: "这是邮件正文".into(),
-            attachments: vec!["report.pdf".into(), "报价.xlsx".into()],
-        };
-
-        let markdown = format_msg_as_markdown(&parts);
-
-        assert!(markdown.contains("发件人: alice@example.com"));
-        assert!(markdown.contains("收件人: bob@example.com"));
-        assert!(markdown.contains("抄送: carol@example.com"));
-        assert!(markdown.contains("密送: audit@example.com"));
-        assert!(markdown.contains("主题: 项目进展"));
-        assert!(markdown.contains("日期: 2026-06-25T10:00:00Z"));
-        assert!(markdown.contains("正文:\n这是邮件正文"));
-        assert!(markdown.contains("附件: report.pdf, 报价.xlsx"));
-    }
-
-    #[test]
-    fn msg_text_cleanup_removes_nul_padding() {
-        assert_eq!(clean_msg_text("OpenAI\0"), "OpenAI");
-        assert_eq!(
-            clean_msg_text("你的临时 OpenAI 登录代码\0"),
-            "你的临时 OpenAI 登录代码"
-        );
-    }
-
-    #[test]
-    fn msg_hex_html_body_decodes_to_readable_text() {
-        let html_hex = "3c68746d6c3e3c686561643e3c7374796c653e2e78207b20636f6c6f723a207265643b207d3c2f7374796c653e3c2f686561643e3c626f64793e3c703e4f70656e414920e799bbe5bd95e4bba3e7a081efbc9a203132333435363c2f703e3c703ee8afb7e58bbfe58886e4baab3c2f703e3c2f626f64793e3c2f68746d6c3e";
-        let text = html_to_text(&decode_msg_html_payload(html_hex));
-
-        assert!(text.contains("OpenAI 登录代码"));
-        assert!(text.contains("123456"));
-        assert!(text.contains("请勿分享"));
-        assert!(!text.contains("3c68746d6c"));
-        assert!(!text.contains("<html>"));
-    }
-
-    #[test]
-    fn msg_sample_from_env_decodes_when_provided() {
-        let Ok(path) = std::env::var("PINVOU3_MSG_SAMPLE") else {
-            return;
-        };
-        let parsed = parse_msg_via_msg_parser(Path::new(&path)).unwrap();
-
-        assert!(!parsed.contains('\0'));
-        assert!(!parsed.contains("3c68746d6c"));
-        assert!(parsed.contains("OpenAI") || parsed.contains("正文:"));
-    }
-
-    #[test]
     fn windows_invalid_msg_returns_warning_without_msgconvert_dependency() {
         if !crate::platform::capabilities::is_windows() {
             return;
@@ -2894,104 +858,6 @@ mod tests {
         assert!(!warning.contains("msgconvert"));
     }
 
-    #[test]
-    fn windows_email_dependency_check_uses_native_msg_parser() {
-        if !crate::platform::capabilities::is_windows() {
-            return;
-        }
-        let deps = check_dependencies();
-        let email = deps
-            .iter()
-            .find(|item| item.key == "email")
-            .expect("email dependency item should exist");
-
-        assert!(email.installed);
-        assert!(email.apt.is_empty());
-        assert!(!email.apt.contains("libemail-outlook-message-perl"));
-        assert!(!email.apt.contains("msgconvert"));
-    }
-
-    #[test]
-    fn eml_regression_parses_headers_body_and_attachment_when_python_available() {
-        if !crate::platform::os::command_exists(&crate::platform::paths::python_command()) {
-            eprintln!("skip: Python is not available");
-            return;
-        }
-
-        let dir =
-            std::env::temp_dir().join(format!("pinvou3-eml-regression-{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let eml = dir.join("m.eml");
-        let raw = concat!(
-            "From: alice@example.com\r\n",
-            "To: bob@example.com\r\n",
-            "Subject: Project Update\r\n",
-            "Date: Thu, 25 Jun 2026 10:00:00 +0800\r\n",
-            "MIME-Version: 1.0\r\n",
-            "Content-Type: multipart/mixed; boundary=\"b\"\r\n\r\n",
-            "--b\r\n",
-            "Content-Type: text/plain; charset=utf-8\r\n\r\n",
-            "This is email body\r\n",
-            "--b\r\n",
-            "Content-Type: text/plain; name=\"note.txt\"\r\n",
-            "Content-Disposition: attachment; filename=\"note.txt\"\r\n\r\n",
-            "attachment\r\n",
-            "--b--\r\n",
-        );
-        std::fs::write(&eml, raw).unwrap();
-
-        let parsed = parse_eml_via_python(&eml).unwrap();
-
-        std::fs::remove_dir_all(&dir).ok();
-        assert!(parsed.contains("alice@example.com"));
-        assert!(parsed.contains("bob@example.com"));
-        assert!(parsed.contains("Project Update"));
-        assert!(parsed.contains("Thu, 25 Jun 2026 10:00:00 +0800"));
-        assert!(parsed.contains("This is email body"));
-        assert!(parsed.contains("note.txt"));
-    }
-
-    #[test]
-    fn pandoc_tool_command_uses_os_layer_program() {
-        let command = pandoc_tool_command();
-        assert_eq!(
-            command.get_program(),
-            crate::platform::os::pandoc_tool_path().as_os_str()
-        );
-    }
-
-    #[test]
-    fn ocr_tool_command_uses_os_layer_program() {
-        let command = ocr_tool_command();
-        assert_eq!(
-            command.get_program(),
-            crate::platform::os::ocr_tool_path().as_os_str()
-        );
-    }
-
-    #[test]
-    fn ocr_tessdata_arg_is_added_when_os_layer_provides_dir() {
-        let mut command = ocr_tool_command();
-        add_ocr_tessdata_arg(&mut command);
-        let args: Vec<_> = command.get_args().map(|arg| arg.to_os_string()).collect();
-        if let Some(dir) = crate::platform::os::ocr_tessdata_dir() {
-            assert!(args.iter().any(|arg| arg == "--tessdata-dir"));
-            assert!(args.iter().any(|arg| arg == dir.as_os_str()));
-        } else {
-            assert!(args.iter().all(|arg| arg != "--tessdata-dir"));
-        }
-    }
-
-    #[test]
-    fn windows_pandoc_missing_message_points_to_repair_install() {
-        if !crate::platform::capabilities::is_windows() {
-            return;
-        }
-        let message = crate::platform::os::pandoc_missing_message();
-        assert!(!message.contains("sudo apt install pandoc"));
-        assert!(message.contains("修复") || message.contains("重新安装"));
-    }
-
     /// 端到端 OCR 实测：依赖本机装了 tesseract + chi_sim，且 /tmp 下有用
     /// PIL 造的中文测试图/扫描件 PDF（见 PR 说明的造图脚本）。常规 CI 无这些
     /// 前置，故 `#[ignore]`；手动 `cargo test -- --ignored ocr_extracts_chinese`。
@@ -2999,7 +865,7 @@ mod tests {
     #[test]
     #[ignore = "需要本机 tesseract+chi_sim 与 /tmp 测试文件"]
     fn ocr_extracts_chinese_from_image_and_scanned_pdf() {
-        if !system_tools().tesseract {
+        if !ingest_deps::system_tools().tesseract {
             eprintln!("跳过：本机无 tesseract");
             return;
         }
@@ -3201,94 +1067,6 @@ mod tests {
 }
 
 #[cfg(test)]
-mod visual_preview_smoke {
-    use super::*;
-    use std::process::Command;
-
-    // 真跑 soffice/pandoc/pdftoppm，验证可视化预览两条路径产出非空 + 图片内联。
-    // 依赖系统工具，CI 无则 ignore。
-    #[test]
-    #[ignore = "需要 libreoffice + pandoc + poppler"]
-    fn office_to_inline_html_and_pdf_to_pngs() {
-        let dir = std::env::temp_dir().join(format!("pinvou3-visual-smoke-{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-
-        // 1) md -> docx (pandoc) -> inline html
-        let md = dir.join("doc.md");
-        std::fs::write(&md, "# 标题\n\n正文一段。\n\n- 列表项\n").unwrap();
-        let docx = dir.join("doc.docx");
-        let ok = Command::new("pandoc")
-            .arg(&md)
-            .arg("-o")
-            .arg(&docx)
-            .status()
-            .map(|s| s.success())
-            .unwrap_or(false);
-        assert!(ok && docx.exists(), "pandoc 造 docx 失败");
-        let html = libreoffice_to_inline_html(&docx).expect("office->html 应成功");
-        assert!(html.contains("标题"), "HTML 应含正文文字");
-        assert!(
-            !html.contains("src=\"doc_html"),
-            "旁置图片应已内联(不应残留相对 src)"
-        );
-
-        // 2) md -> pdf (soffice via docx) -> png 页
-        let pdf = dir.join("doc.pdf");
-        let _ = libreoffice_tool_command()
-            .arg(libreoffice_user_installation_arg(&dir.join("p")).unwrap())
-            .args(["--headless", "--convert-to", "pdf", "--outdir"])
-            .arg(&dir)
-            .arg(&docx)
-            .status();
-        if pdf.exists() {
-            let (imgs, _trunc) = pdf_to_png_data_uris(&pdf, 30).expect("pdf->png 应成功");
-            assert!(!imgs.is_empty(), "应产出至少一页");
-            assert!(
-                imgs[0].starts_with("data:image/png;base64,"),
-                "应为 png data URI"
-            );
-        }
-
-        // 3) md -> pptx (pandoc) -> office_to_png(演示稿走 PDF→PNG)
-        let pptx = dir.join("deck.pptx");
-        if Command::new("pandoc")
-            .arg(&md)
-            .arg("-o")
-            .arg(&pptx)
-            .status()
-            .map(|s| s.success())
-            .unwrap_or(false)
-        {
-            let (imgs, _t) = office_to_png_data_uris(&pptx, 30).expect("pptx->png 应成功");
-            assert!(
-                !imgs.is_empty() && imgs[0].starts_with("data:image/png;base64,"),
-                "pptx 应产出页图"
-            );
-        }
-
-        // 4) csv -> xlsx (soffice) -> inline html(电子表格走 HTML 表格)
-        let csv = dir.join("data.csv");
-        std::fs::write(&csv, "甲,乙\n1,2\n3,4\n").unwrap();
-        let _ = libreoffice_tool_command()
-            .arg(libreoffice_user_installation_arg(&dir.join("p2")).unwrap())
-            .args(["--headless", "--convert-to", "xlsx", "--outdir"])
-            .arg(&dir)
-            .arg(&csv)
-            .status();
-        let xlsx = dir.join("data.xlsx");
-        if xlsx.exists() {
-            let html = libreoffice_to_inline_html(&xlsx).expect("xlsx->html 应成功");
-            assert!(
-                html.contains('甲') || html.to_lowercase().contains("table"),
-                "xlsx HTML 应含表格内容"
-            );
-        }
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-}
-
-#[cfg(test)]
 mod multi_type_dispatch_e2e {
     use super::*;
 
@@ -3306,7 +1084,7 @@ mod multi_type_dispatch_e2e {
     #[test]
     #[ignore = "需要 pandoc + pdftotext + libreoffice + tesseract(本机齐全时跑)"]
     fn ingest_dispatches_each_real_type_by_extension_not_binary() {
-        let tools = system_tools();
+        let tools = ingest_deps::system_tools();
         let dir = std::env::temp_dir().join(format!(
             "pinvou3-multi-type-e2e-{}-{}",
             std::process::id(),
